Add tests and Dokken support
We test as many aspects as possible of the installed supervisord service 
and its support binaries. We've also switched to Dokken when possible 
because it tests things a little faster than Vagrant.

Vagrant tests include a partial Vagrantfile designed to speed up testing 
in Virtualbox.

Added a test recipe to use for testing.